Create MSI Repair
A quick and dirty MSI server repair. Sets the service to on demand startup, then stops the service, then unregisters, and reregisters the regserver components, verifies and points to correct components then restarts the service. Save as a quick .bat or .cmd file and set up as a startup task on those pesky malware infections