Create DIsableStartupServicesREGADD
This is a REAL dirty way of using an elevated command prompt CMD to copy paste or run as a .cmd (ran as admin) to add the registry DWORD value of 4 to the listed services. The value of 4 goes under a value of startup type, 4 - startuptype= Disabled. Thus anything added in the same manner as shown above (service wise) can be set to startup type disabled in worst case scenarios where your usual services.msc settings menu is not working.