Create Automated-ABTZ-Security-Baseline.ps1
This script is to be used via powershell, and is designed around the idea of creating an automated cybersecurity baseline installer script. This will handle the brunt of the work, the rest is up to the user. Though the code is clearly labeled, please take a moment to test the code in a test environment before running it on your windows OS installs. Please note this is specifically tailored to windows 10 home x64, and would have to be altered to work safely with any other windows install. I do not take responsibility for any damage caused by misuse or misunderstanding the intention and functionality of the script.