Create MaxSteel.ps1
MaxSteel.ps1 v1.0 is an attempt to provide the utmost hardening potential from an initial win10 home OS install using only powershell. This script may not cover all potential hardening avenues and may leave gals in the attack surface, so please, treat it as an automated baseline to build from. As always, please test this script in a test environment first before deploying it to your system. Please be aware this is for a windows 10 home install of the 2023 variant. Using this on any other version such as windows 10 pro, WILL BREAK THINGS.  I am not responsible for the abuse or misuse of this script.